rename: user 관련 요청 dto 파일 위치 이동 #11
