+++ src/main/java/com/second/kirby/dto/request/user/LoginRequest.java
@@ -1,4 +1,4 @@
-package com.second.kirby.dto.request;
+package com.second.kirby.dto.request.user;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import jakarta.validation.constraints.NotBlank;
